Refactor HomePage and QuizPage components to utilize new BackgroundImage and QuizSelector components for improved structure and maintainability
